pyproject.toml
[project]
name = "headease-homemonitoring-poc"
version = "0.1.0"
description = "HeadEase Home Monitoring PoC - Data Holder for GF Proeftuin"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "cryptography>=44",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.20",
    "pysodium>=0.7",
    "PyJWT>=2.10",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8",
    "pytest-asyncio>=0.25",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
